On branch master  Changes to be committed:    (use "git reset HEAD <file>..." to unstage)
	new file:   CodeBook_GCD_Project.Rmd
	new file:   HARdataSummary.txt
	new file:   run_analysis.R

These are the three files needed for the class project for Getting and Cleaning Data, Coursera.  The Code Book describes
the data, where it came from and what it means including units.  The HARdataSummary.txt is the modified data set itself.
The run_analysis.R file is the R code that will reproduce the HARdataSummary.txt file.